change to move some helper functions into Olivia's working structure, adjust import statements for clarity
