+++ scripts/analysis_comprehensive_TLS_v0.py
@@ -2,12 +2,14 @@
 import sys
 sys.path.append('../src/qicklab/analysis')
 
-from analysis_helper_TLS_comprehensive import *
+from analysis_helper_TLS_comprehensive import qspec, t1, ssf, resstarkspec, starkspec, auto_threshold
 
 ############### set values here ###################
-QubitIndex = 0 #zero indexed
-data_dir = "/Users/joycecs/PycharmProjects/PythonProject/.venv/QUIET/QUIET_data/RR_comprehensive_TLS/" #update based on file transfer location from Ryan
+# data_dir = "/Users/joycecs/PycharmProjects/PythonProject/.venv/QUIET/QUIET_data/RR_comprehensive_TLS/" #update based on file transfer location from Ryan
+data_dir = "/exp/cosmiq/data/QUIET/QICK_data/run6/6transmon/TLS_Comprehensive_Study/source_off_substudy1/" #update based on file transfer location from Ryan
 dataset = '2025-04-15_21-24-46'
+
+QubitIndex = 0 #zero indexed
 analysis_flags = {"get_threshold": True, "load_all_data": True, "timestream": True, "round": True}
 selected_round = [10, 73]
 threshold = 0 #overwritten when get_threshold flag is set to True
